requirements.txt
Flask==3.0.0
pypdf==3.17.4
groq==0.4.1
python-dotenv==1.0.0
gunicorn==21.2.0